requirements.txt
ConfigParser==7.2.0
cryptography==45.0.3
Cython==3.1.1
dl==0.1.0
HTMLParser==0.0.2
ipaddr==2.2.0
lxml==5.4.0
matplotlib==3.9.1
numpy==1.23.5
ordereddict==1.1
Pillow==11.2.1
protobuf==6.31.1
pyOpenSSL==25.1.0
redis==6.2.0
thread==2.0.5
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
wincertstore==0.2.1
